requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1
pandas
numpy
PyYAML